requirements.txt
# x-generator dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
supabase>=2.3.0

# LLM clients
openai>=1.10.0
anthropic>=0.18.0

# RAG / Vector store
chromadb

# HTTP client
httpx>=0.26.0

# Scheduler
apscheduler>=3.10.4

# OAuth2
authlib>=1.3.0
python-jose[cryptography]>=3.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Timezone handling
pytz>=2024.1
